type initial props as any
